Started to shift to an annotation-based system
Should allow for more powerful load/save/etc, plans are to allow people
to just mark what params need to be read in/out of by the system, and
what types of data they should be. As a furthur result, this may also
allow for shifting away from needing specific exportAsJSON methods, as
well as removing boilerplate code for decoding json into the internal
format, etc. This is getting smarter!!